Rename kernel library and physical memory manager directories.
