+++ os/kernel/src/interrupts/syscalls/handlers/memory_calls.h
@@ -1,7 +1,7 @@
 #ifndef MEMORY_CALLS_H
 #define MEMORY_CALLS_H
 
-#include "memory/manager/physic/physical_memory_manager.h"
+#include "memory/manager/physical/physical_memory_manager.h"
 
 void syscall_memory_get_physical_memory_stats(interrupt_state *state);
