+++ os/kernel/src/klibrary/ktime.c
