+++ os/kernel/src/klibrary/kvector.c
