+++ os/kernel/src/klibrary/kvector.h
