+++ os/kernel/src/memory/manager/virtual/virtual_memory_manager.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include "memory/paging/paging.h"
-#include "memory/manager/physic/physical_memory_manager.h"
+#include "memory/manager/physical/physical_memory_manager.h"
 
 uint32_t virtual_memory_alloc_page(bool supervisor);
 bool virtual_memory_dealloc_page(uint32_t index);
